feat: adding gb poisoner class and updating code
